// src/models/bang-diem-tong/dto/create-bang-diem-tong.dto.ts
export class CreateBangDiemTongDto {
    hocSinh: string;
    GVCN: string;
    bangDiemMon: string[];
    nhanXet: string;
    xepLoai: string;

    hanhKiem_hk1: string;
    hocLuc_hk1: string;
    diemTB_hk1: number;

    hanhKiem_hk2: string;
    hocLuc_hk2: string;
    diemTB_hk2: number;

    hanhKiem_caNam: string;
    hocLuc_caNam: string;
    diemTB_caNam: number;
}